Enhance structure of load generator (#42)
* enhance(load-gen): move function-specific config, setup and executor to seperate files to increase modularity
- improve documentation
- rearrange a bunch of stuff

* fix: rearrange so that it actually works

* Clean up unused files

* Move helper functions to utils.js

* Move metrics to metrics.js

* docs: improve documentation of load generator

---------
