--- load-generator/config/config.json
@@ -1,18 +1,0 @@
-{
-    "metaData": {
-        "paramFile": "params.csv",
-        "protoFiles": [
-            "definitions.proto",
-            "common.proto"
-        ]
-    },
-    "function": {
-        "name": "MyFunction",
-        "serviceFn": "test.MyService/MyFunction",
-        "params": [
-            "param1",
-            "param2",
-            "..."
-        ]
-    }
-}
