--- load-generator/config/params.csv
@@ -1,4 +1,0 @@
-seconds,vus,rps,param1,param2,...
-2,40,20,20,10
-3,60,30,30,10
-...
